+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.0-SNAPSHOT (10.3.2017)
+## 0.2.0-20170315.061825-12 (15.3.2017)
 
 * New option to allow empty input on decode, `:allow-empty-input-on-decode?` (default to `false`). If set to `true`, empty inputstreams map to `nil` body, otherwise, the decoder decides what happens (transit fails on default).
   * Fixes [#33](https://github.com/metosin/muuntaja/issues/33)